motion detector
Detects motion of objects ,So that you monitor place ,and know when people enter your door step. Compares the current frame and the previouse frame together inorder to detect motion. It can differ between the motion of a human and the slight motion caused by winds helping it to react only if a human passes by. Press t to start and press e to exit.